aject works ok

+++ app/controllers/wco_email/email_filters_controller.rb
@@ -5,25 +5,43 @@ class WcoEmail::EmailFiltersController < WcoEmail::ApplicationController
 
   def create
     authorize! :create, WcoEmail::EmailFilter
-    @email_filter = WcoEmail::EmailFilter.create params[:email_filter].permit!
-    if @email_filter.persisted?
+    @email_filter = WcoEmail::EmailFilter.new
+
+    ## conditions, skip_conditions aren't wired for aject yet.
+    (params[:email_filter][:actions_attributes]||[]).each do |_key, attrs|
+      type, id = attrs.delete(:aject).split(' ')
+      attrs[:aject_type] = type
+      attrs[:aject_id] = id
+    end
+
+    flag = @email_filter.update_attributes( params[:email_filter].permit! )
+
+    if flag
       flash[:notice] = 'Success'
+      redirect_to action: :show, id: @email_filter.id
     else
       flash[:alert] = "No luck: #{@email_filter.errors.full_messages.join(', ')}."
+      redirect_to request.referrer
     end
-    redirect_to action: 'new'
+
   end
 
   def destroy
-    @email_filter = WcoEmail::EmailFilter.find params[:id]
-    authorize! :destroy, @email_filter
-    flag = @email_filter.destroy
+    authorize! :destroy, WcoEmail::EmailFilter
+
+    if params[:id]
+      @email_filter = WcoEmail::EmailFilter.find params[:id]
+      flag = @email_filter.destroy
+    elsif params[:ids]
+      flag = WcoEmail::EmailFilter.find( params[:ids] ).map { |ef| ef.destroy }
+    end
+
     if flag
       flash[:notice] = 'Success'
     else
       flash[:alert] = 'Error'
     end
-    redirect_to action: 'index'
+    redirect_to request.referrer
   end
 
   def edit
@@ -34,7 +52,6 @@ class WcoEmail::EmailFiltersController < WcoEmail::ApplicationController
     @email_filter = WcoEmail::EmailFilter.find( params[:id] )
     authorize! :edit, @email_filter
 
-    # @new_email_filter_condition = WcoEmail::EmailFilterCondition.new
     @email_filter.conditions.build
     @email_filter.skip_conditions.build
     @email_filter.actions.build
@@ -43,6 +60,7 @@ class WcoEmail::EmailFiltersController < WcoEmail::ApplicationController
       'none' => [ [nil,nil] ],
       'Wco::Tag' => Wco::Tag.all.map { |t| [t.slug, "Wco::Tag #{t.id}" ] },
       'WcoEmail::EmailTemplate' => WcoEmail::EmailTemplate.all.map { |t| [ t.slug, "WcoEmail::EmailTemplate #{t.id}" ] },
+      'WcoEmail::EmailActionTemplate' => WcoEmail::EmailActionTemplate.all.map { |t| [ t.slug, "WcoEmail::EmailActionTemplate #{t.id}" ] },
     }
   end
 
@@ -50,10 +68,12 @@ class WcoEmail::EmailFiltersController < WcoEmail::ApplicationController
   def index
     authorize! :index, WcoEmail::EmailFilter.new
     @email_filter  = WcoEmail::EmailFilter.new
-    @email_filters = WcoEmail::EmailFilter.all().includes( :email_template, :conversations )
+    @email_filters = WcoEmail::EmailFilter.all.includes( :email_template, :conversations )
 
     if params[:q]
       @email_filters = @email_filters.where( from_exact: /#{params[:q]}/i )
+    else
+      @email_filters = @email_filters.active
     end
 
     @email_filters = @email_filters.page( params[WcoEmail::EmailFilter::PAGE_PARAM_NAME]
@@ -68,17 +88,66 @@ class WcoEmail::EmailFiltersController < WcoEmail::ApplicationController
     @email_filter = WcoEmail::EmailFilter.new
     authorize! :new, @email_filter
 
-    @new_email_filter_condition = WcoEmail::EmailFilterCondition.new
+    @email_filter.conditions.build
+    @email_filter.skip_conditions.build
+    @email_filter.actions.build
+
+    @aject_options = {
+      'none' => [ [nil,nil] ],
+      'Wco::Tag' => Wco::Tag.all.map { |t| [t.slug, "Wco::Tag #{t.id}" ] },
+      'WcoEmail::EmailTemplate' => WcoEmail::EmailTemplate.all.map { |t| [ t.slug, "WcoEmail::EmailTemplate #{t.id}" ] },
+    }
   end
 
   def show
     @email_filter = WcoEmail::EmailFilter.find params[:id]
     authorize! :show, @email_filter
+    @conversations = @email_filter.conversations.page( params[WcoEmail::Conversation::PAGE_PARAM_NAME] ).per( current_profile.per_page )
   end
 
   def update
     @email_filter = WcoEmail::EmailFilter.find params[:id]
     authorize! :update, @email_filter
+
+    ## conditions, skip_conditions aren't wired for aject yet.
+    (params[:email_filter][:actions_attributes]||[]).each do |_key, attrs|
+      type, id = attrs.delete(:aject).split(' ')
+      attrs[:aject_type] = type
+      attrs[:aject_id] = id
+    end
+
+    flag = @email_filter.update_attributes( params[:email_filter].permit! )
+
+    if flag
+      flash[:notice] = 'Success'
+      redirect_to request.referrer
+    else
+      flash[:alert] = "No luck: #{@email_filter.errors.full_messages.join(', ')}."
+      redirect_to request.referrer
+    end
+
+  end
+
+  ##
+  ## private
+  ##
+  private
+
+  def set_lists
+    @tags_list = Wco::Tag.list
+    @email_templates_list        = WcoEmail::EmailTemplate.list
+    @email_actions_list          = WcoEmail::EmailAction.list
+    @email_action_templates_list = WcoEmail::EmailActionTemplate.list
+  end
+
+
+
+end
+
+
+
+
+
 
 =begin
     params[:email_filter][:actions_attributes].each do |_key, attrs|
@@ -109,42 +178,3 @@ class WcoEmail::EmailFiltersController < WcoEmail::ApplicationController
 
     end
 =end
-
-    params[:email_filter][:actions_attributes].each do |_key, attrs|
-      type, id = attrs.delete(:aject).split(' ')
-      attrs[:aject_type] = type
-      attrs[:aject_id] = id
-    end
-
-    # if params[:email_filter][:tag].blank?
-    #   params[:email_filter].delete :tag
-    # end
-
-    flag = @email_filter.update_attributes( params[:email_filter].permit! )
-
-    if flag
-      flash[:notice] = 'Success'
-      redirect_to request.referrer
-    else
-      flash[:alert] = "No luck: #{@email_filter.errors.full_messages.join(', ')}."
-      redirect_to request.referrer
-    end
-
-  end
-
-  ##
-  ## private
-  ##
-  private
-
-  def set_lists
-    @tags_list = Wco::Tag.list
-    @email_templates_list        = WcoEmail::EmailTemplate.list
-    @email_actions_list          = WcoEmail::EmailAction.list
-    @email_action_templates_list = WcoEmail::EmailActionTemplate.list
-  end
-
-
-
-end
-
